requirements.txt
Jinja2-2.8
MarkupSafe-0.23
Werkzeug-0.10.4
flask-0.10.1
itsdangerous-0.24
blinker-1.4
flask-mail-0.9.1
SQLAlchemy-1.0.8
flask-SQLAlchemy-2.0
flask-login-0.3.0
WTForms-2.0.2
flask-wtf-0.12
dominate-2.1.16
flask-bootstrap-3.3.5.6
flask-migrate==1.6.0
flask-script==2.0.5
flask-material=0.1.1
flask-moment=0.5.1